Michael could you please stop murdering people

+++ db/migrate/20240613235709_create_quiz_undertakings.rb
@@ -4,7 +4,7 @@ class CreateQuizUndertakings < ActiveRecord::Migration[7.1]
       t.belongs_to :account, null: false, foreign_key: true
       t.belongs_to :quiz, null: false, foreign_key: true
       
-      t.integer :correct_answers, null: false, default: 0
+      t.float :correct_answers, null: false, default: 0
 
       t.timestamps
     end
